requirements.txt
# Data Handling
pandas
numpy
tqdm               # Progress bars (crucial for large loops)
pillow             # Image processing

# Deep Learning (The "Brains")
torch              # PyTorch
torchvision        # For ResNet50
transformers       # For text models
sentence-transformers # The industry standard for text embeddings

# Vector Search (Fast similarity search on GPU if available)
faiss-cpu          # Facebook AI Similarity Search (GPU version)
# Note: If you are on Windows, use 'faiss-cpu' if 'faiss-gpu' fails to build, 
# or use WSL2. For now, try faiss-gpu.
# --- For the Conversational "G" in RAG ---
accelerate     # Required for running models efficiently
bitsandbytes   # For 8-bit/4-bit quantization (to save VRAM)